AG-13178 - Make line with text util functions instead of class
