+++ packages/ag-charts-enterprise/src/features/annotations/disjoint-channel/disjointChannelScene.ts
@@ -5,8 +5,8 @@ import { AnnotationScene } from '../scenes/annotationScene';
 import { ChannelScene } from '../scenes/channelScene';
 import { CollidableText } from '../scenes/collidableTextScene';
 import { DivariantHandle, UnivariantHandle } from '../scenes/handle';
-import { LineWithTextScene } from '../scenes/lineWithTextScene';
-import { convertPoint, invertCoords } from '../utils/values';
+import { updateChannelText } from '../utils/lineWithText';
+import { invertCoords } from '../utils/values';
 import type { DisjointChannelProperties } from './disjointChannelProperties';
 
 const { Vec2, Vec4 } = _ModuleSupport;
@@ -165,15 +165,7 @@ export class DisjointChannelScene extends ChannelScene<DisjointChannelProperties
     updateText(datum: DisjointChannelProperties, top: _ModuleSupport.Vec4, bottom: _ModuleSupport.Vec4) {
         this.text = this.updateNode(CollidableText, this.text, !!datum.text.label);
 
-        LineWithTextScene.updateChannelText(
-            false,
-            top,
-            bottom,
-            datum.text,
-            datum.strokeWidth,
-            this.text,
-            datum.text.label
-        );
+        updateChannelText(false, top, bottom, datum.text, datum.strokeWidth, this.text, datum.text.label);
     }
 
     override getBackgroundPoints(
