+++ packages/ag-charts-enterprise/src/features/annotations/parallel-channel/parallelChannelScene.ts
@@ -6,7 +6,7 @@ import { ChannelScene } from '../scenes/channelScene';
 import { CollidableLine } from '../scenes/collidableLineScene';
 import { CollidableText } from '../scenes/collidableTextScene';
 import { DivariantHandle, UnivariantHandle } from '../scenes/handle';
-import { LineWithTextScene } from '../scenes/lineWithTextScene';
+import { updateChannelText } from '../utils/lineWithText';
 import { isPoint, validateDatumPoint } from '../utils/validation';
 import { invertCoords } from '../utils/values';
 import type { ParallelChannelProperties } from './parallelChannelProperties';
@@ -215,15 +215,7 @@ export class ParallelChannelScene extends ChannelScene<ParallelChannelProperties
     updateText(datum: ParallelChannelProperties, top: _ModuleSupport.Vec4, bottom: _ModuleSupport.Vec4) {
         this.text = this.updateNode(CollidableText, this.text, !!datum.text.label);
 
-        LineWithTextScene.updateChannelText(
-            true,
-            top,
-            bottom,
-            datum.text,
-            datum.strokeWidth,
-            this.text,
-            datum.text.label
-        );
+        updateChannelText(true, top, bottom, datum.text, datum.strokeWidth, this.text, datum.text.label);
     }
 
     override getBackgroundPoints(
